Show peak object instances on the Caller/Callee tab
Add a new column to the table to display the peak number of
object instances for a given location.

The idea is to make it easier to detect an unexpected growth
in number of objects. One example is when a buggy cache
implementation is not evicting entries. The objects are
referenced and will be destructed on exit (thus not leaking
in the classical sense) but still using memory and growing
memory usage during the process execution.

Errors like this are hard to spot by analyzing memory usage
patterns if the size of the object is small, but sometimes
easy by looking at peak number of instances.